Modified abstractions to facilitate introduction of `GivenEachOf`.
The builder API is unchanged, but necessary changes to the VSTest adapter have also been made. Namely, the "[Test Arrangement]" result of each test is gone, and when test arrangement fails, you no longer get each assertion listed as failed. This is because in theory we can now have multiple test cases - and if the arrangement fails, we don't know how many there are - so can't list the assertions for each of them.