BREAKING CHANGE: Split "Then" into "Then" and "ThenThrows"..
The when clause of a test will either complete successfully or throw. As such, it doesn't make much sense for assertions to deal with an object that gives access to both the return value and the exception - and this approach makes for needlessly long assertions. So, going forwards, we have "Then" that (fails by throwing the exception if one was thrown) and (in the case of test functions) gives you just the return value to work with; and "ThenThrows" for expecting an exception.

The only situation this perhaps makes a little more awkward is if all you care about is the When clause NOT throwing (which was previously easy by asserting that the outome's exception is null) - but this isn't a common scenario, and could always be covered with a "ThenDoesntThrow".